Basic working keras wrapper
