+++ credoai/artifacts/__init__.py
@@ -8,7 +8,9 @@ them.
 from .data.base_data import Data
 from .data.comparison_data import ComparisonData
 from .data.tabular_data import TabularData
+from .data.tensor_data import TensorData
 from .model.base_model import Model
 from .model.classification_model import ClassificationModel, DummyClassifier
 from .model.comparison_model import ComparisonModel, DummyComparisonModel
 from .model.regression_model import DummyRegression, RegressionModel
+from .model.keras_classification_model import KerasClassificationModel
